fixup: abstract spawn
This can be shared between the callback + promise API via utils

Slight concern that moving this code to internals could break
monkey patching.
