+++ lib/child_process.js
@@ -24,7 +24,6 @@
 const util = require('util');
 const { getSystemErrorName } = require('internal/util');
 const { isArrayBufferView } = require('internal/util/types');
-const debug = util.debuglog('child_process');
 const { Buffer } = require('buffer');
 const { Pipe, constants: PipeConstants } = internalBinding('pipe_wrap');
 const {
@@ -41,9 +40,11 @@ const {
   ChildProcess
 } = child_process;
 const {
+  debug,
   normalizeExecArgs,
   normalizeSpawnArguments,
   sanitizeKillSignal,
+  spawn,
   validateMaxBuffer,
   validateTimeout
 } = require('internal/child_process/utils');
@@ -372,28 +373,7 @@ Object.defineProperty(exports.execFile, util.promisify.custom, {
 });
 
 
-var spawn = exports.spawn = function spawn(/* file, args, options */) {
-  var opts = normalizeSpawnArguments.apply(null, arguments);
-  var options = opts.options;
-  var child = new ChildProcess();
-
-  debug('spawn', opts.args, options);
-
-  child.spawn({
-    file: opts.file,
-    args: opts.args,
-    cwd: options.cwd,
-    windowsHide: !!options.windowsHide,
-    windowsVerbatimArguments: !!options.windowsVerbatimArguments,
-    detached: !!options.detached,
-    envPairs: opts.envPairs,
-    stdio: options.stdio,
-    uid: options.uid,
-    gid: options.gid
-  });
-
-  return child;
-};
+exports.spawn = spawn;
 
 function spawnSync(/* file, args, options */) {
   var opts = normalizeSpawnArguments.apply(null, arguments);
