+++ lib/internal/child_process/promises.js
@@ -6,42 +6,13 @@ const { Buffer } = require('buffer');
 const {
   ERR_CHILD_PROCESS_STDIO_MAXBUFFER,
 } = require('internal/errors').codes;
-const child_process = require('internal/child_process/base');
-const {
-  ChildProcess
-} = child_process;
 const {
   normalizeExecArgs,
-  normalizeSpawnArguments,
   sanitizeKillSignal,
+  spawn,
   validateMaxBuffer,
   validateTimeout
 } = require('internal/child_process/utils');
-
-const debug = util.debuglog('child_process');
-
-function spawn(/* file, args, options */) {
-  var opts = normalizeSpawnArguments.apply(null, arguments);
-  var options = opts.options;
-  var child = new ChildProcess();
-
-  debug('spawn', opts.args, options);
-
-  child.spawn({
-    file: opts.file,
-    args: opts.args,
-    cwd: options.cwd,
-    windowsHide: !!options.windowsHide,
-    windowsVerbatimArguments: !!options.windowsVerbatimArguments,
-    detached: !!options.detached,
-    envPairs: opts.envPairs,
-    stdio: options.stdio,
-    uid: options.uid,
-    gid: options.gid
-  });
-
-  return child;
-}
 
 async function exec() {
   const opts = normalizeExecArgs.apply(null, arguments);
