+++ lib/internal/child_process/utils.js
@@ -1,6 +1,9 @@
 'use strict';
-
+const { debuglog } = require('util');
+const debug = debuglog('child_process');
 const { convertToValidSignal, deprecate } = require('internal/util');
+const { ChildProcess } = require('internal/child_process/base');
+
 const {
   ERR_INVALID_ARG_VALUE,
   ERR_INVALID_ARG_TYPE,
@@ -200,10 +203,35 @@ function normalizeSpawnArguments(file, args, options) {
   };
 }
 
+function spawn(/* file, args, options */) {
+  var opts = normalizeSpawnArguments.apply(null, arguments);
+  var options = opts.options;
+  var child = new ChildProcess();
+
+  debug('spawn', opts.args, options);
+
+  child.spawn({
+    file: opts.file,
+    args: opts.args,
+    cwd: options.cwd,
+    windowsHide: !!options.windowsHide,
+    windowsVerbatimArguments: !!options.windowsVerbatimArguments,
+    detached: !!options.detached,
+    envPairs: opts.envPairs,
+    stdio: options.stdio,
+    uid: options.uid,
+    gid: options.gid
+  });
+
+  return child;
+}
+
 module.exports = {
+  debug,
   normalizeExecArgs,
   normalizeSpawnArguments,
   sanitizeKillSignal,
+  spawn,
   validateTimeout,
   validateMaxBuffer
 };
